Store call stack addresses at pointer width
DebugHelp::getCallStack() handed the caller's uint32 buffer straight to
backtrace(), which writes void* entries. Under LP64 that wrote 8 bytes per
frame into a 4-byte-per-frame array, overrunning every buffer by 2x:

  Fatal.cpp        uint32[68] = 272 bytes, up to 544 written (stack)
  CallStack        uint32[32] = 128 bytes, 208 written (heap)
  CallStackCollector uint32[26] = 104 bytes, 208 written (heap)

Fatal.cpp's formatMessage() backs every WARNING and FATAL, so any warning
that captured a stack corrupted memory. Observed as SIGSEGV inside cfree()
with si_addr 0x8 after an unrelated warning fired.

Capture into a native void* array and widen into a uint64 buffer, which is
correct under both ILP32 and LP64, and widen the storage and printing to
match. lookupAddress() already took uint64 on linux; win32 now agrees and no
longer truncates StackWalk's DWORD64 address to 32 bits.

Also fixes CallStackCollector copying only one element instead of the whole
call stack, and CRCing only half the buffer.

+++ engine/shared/library/sharedDebug/src/linux/DebugHelp.h
@@ -19,7 +19,10 @@ public:
 	static void install();
 	static void remove();
 
-	static void getCallStack(uint32 *callStack, int sizeOfCallStack);
+	//-- Entries are instruction addresses, so the buffer must be pointer-width
+	//   capable. uint64 matches lookupAddress() below and is wide enough on
+	//   both ILP32 and LP64.
+	static void getCallStack(uint64 *callStack, int sizeOfCallStack);
 	static bool lookupAddress(uint64 address, char *libName, char *fileName, int fileNameLength, int &line);
 };
